Adding rendering to the templates with yaml and data working
