+++ src/node/helpers/render-head-assets.ts
@@ -1,0 +1,12 @@
+import {Bundle} from '../create-bundle';
+
+export function renderHeadAssets(bundle: Bundle) {
+  /* const lines = [];
+  for (const inlineStyle of compilation.styles.inline) {
+    lines.push(`<style>${handlebars.escapeExpression(inlineStyle.trim())}</style>`);
+  }
+  for (const syncStyle of compilation.styles.sync) {
+    lines.push(`<link rel="stylesheet" type="text/css" href="${handlebars.escapeExpression(syncStyle)}" />`);
+  }
+  return new handlebars.SafeString(lines.join('\n'));*/
+}
