+++ test/static/extra-files/partials-example-import.tmpl
@@ -23,5 +23,10 @@ scripts:
 partials:
   - ./partials-example-nested-import.tmpl
 ---
+<h2>HTML</h2>
+## MD
+
+{{ yaml.hello }}
+{{ data.hello }}
 
 {{> ./partials-example-nested-import.tmpl}}
