+++ test/static/extra-files/partials-example-nested-import.tmpl
@@ -21,6 +21,8 @@ scripts:
   async:
     - ./scripts/async-3.js
 ---
-<p>Nested partial HTML</p>
+<h3>HTML</h3>
+### MD
 
-Nest partial MD
+{{ yaml.hello }}
+{{ data.hello }}
